requirements.txt
dearpygui==1.10.1
pywin32
numpy
pillow